pyproject.toml
[tool.poetry]
name = "damai-ticket-automation"
version = "0.1.0"
description = "Automated ticket purchasing system for Damai.com"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [
    { include = "web" },
    { include = "mobile" },
    { include = "shared" }
]

[tool.poetry.dependencies]
python = "^3.8"
selenium = ">=4.22.0,<4.28"
appium-python-client = "^4.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^5.0.0"
pytest-mock = "^3.14.0"

[tool.poetry.scripts]
test = "pytest:main"
tests = "pytest:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov=web",
    "--cov=mobile",
    "--cov-branch",
    "--cov-report=term-missing:skip-covered",
    "--cov-report=html:htmlcov",
    "--cov-report=xml:coverage.xml",
    "--cov-fail-under=84",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Tests that take a long time to run",
]

[tool.coverage.run]
source = ["web", "mobile"]
omit = [
    "*/tests/*",
    "*/__init__.py",
    "*/conftest.py",
    "web/quick_diagnosis.py",
    "web/logger.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if __name__ == .__main__.:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
precision = 2
show_missing = true
skip_covered = true

[tool.coverage.html]
directory = "htmlcov"

[tool.coverage.xml]
output = "coverage.xml"